+++ doc/source/changelog.rst
@@ -47,6 +47,8 @@ Python
 * Remove :py:mod:`qibuild.archive` use :py:mod:`qisys.archive` instead.
 * Add a function :py:func:`qibuild.toc.get_build_folder_name` to get the name of
   the build directory.
+* Remove :py:func:`qibuild.toc.set_build_folder_name` and
+  ``qibuild.toc.Toc.build_folder_name``
 
 V2.0
 ----